Create Cause_of_death_analysis_UT.py
This program uses the 'NCHS_-_Leading_Cause_of_Death__United_States.csv' for its analysis which can be found along side this code file. 
You can also view the graphs in this code under the seaborn styling if the seaborn package is installed. This can be installed with 'pip install seaborn' and 
removing the hashed commands.